more work in admin page

+++ modules/file-manager.js
@@ -40,7 +40,7 @@ const startWork = (model) => {
   let data = { id: model.Id, dir: model.FullPath };
   worker.send(data);
 };
-
+// List all hdd
 module.exports.diskLoader = () => {
   drivelist.list().then((drives) => {
     let disks = [];
@@ -62,7 +62,7 @@ module.exports.diskLoader = () => {
     socket.emit("disk-loaded", disks);
   });
 };
-
+//Load content of a folder
 module.exports.loadContent = (data) => {
   //If is it root of disk return;
   if (fs.existsSync(data.Path)) {
@@ -79,7 +79,7 @@ module.exports.loadContent = (data) => {
     socket.emit("content-loaded", { data: tdata, Id: data.Id });
   }
 };
-
+//Scan all files of a direcotry
 module.exports.scanDir = async ({ Id, Path }) => {
   //If is it root of disk return;
   if (!Id && !Path) return socket.emit("scan-info", "Id And Path both can't be null");
@@ -126,16 +126,47 @@ module.exports.scanDir = async ({ Id, Path }) => {
   socket.emit("scan-info", msg);
   console.log(msg);
 };
+/****************** Rename File *******************/
+module.exports.renameFile = async ({ Id, Name }) => {
+  let file = await db.file.findByPk(Id);
+  let success = false;
+  let msg = "File was not found";
+  if (file) {
+    let fromFile = path.join(file.FullPath, file.Name);
+    let toFile = path.join(file.FullPath, Name);
 
-/************Remove file */
+    let fromCover = path.join("./images", file.Cover);
+    let toCover = path.join(path.dirname(fromCover), Name + ".jpg");
+
+    try {
+      if (fs.existsSync(fromFile)) {
+        fs.moveSync(fromFile, toFile);
+        await db.file.update({ Name, Type: file.Type, Cover: "" }, { where: { Id } });
+        success = true;
+        if (fs.existsSync(fromCover)) {
+          fs.moveSync(fromCover, toCover);
+        }
+        msg = `File ${file.Name} was rename to ${Name} successfully`;
+      } else {
+        msg = `File ${file.Name} not found on server`;
+      }
+    } catch (err) {
+      console.log(err);
+    }
+  } else {
+    msg = "File not found on db";
+  }
+  socket.emit("file-renamed", { success, msg, Name });
+};
+/************ Remove file from db and system ***********************/
 
 module.exports.removeFile = async ({ Id, Del }) => {
   let file = await db.file.findByPk(Id);
-  const message = { removed: false, msg: "" };
+  const message = { success: false, msg: "" };
   if (file) {
     try {
       await file.destroy();
-      message.removed = true;
+      message.success = true;
       if (Del) {
         let cover = path.join(
           process.cwd(),
@@ -144,9 +175,14 @@ module.exports.removeFile = async ({ Id, Del }) => {
           file.Type,
           file.Name + ".jpg"
         );
-        fs.removeSync(cover);
-        fs.removeSync(path.join(file.FullPath, file.Name));
-        message.msg = `File ${file.Name} removed from server`;
+        if (fs.existsSync(cover)) fs.removeSync(cover);
+        let fPath = path.join(file.FullPath, file.Name);
+        if (fs.existsSync(fPath)) {
+          fs.removeSync(fPath);
+          message.msg = `File ${file.Name} removed from server`;
+        } else {
+          message.msg = `File Don't exit on server was only remove from db`;
+        }
       } else {
         message.msg = `File ${file.Name} removed from DB`;
       }
@@ -155,7 +191,59 @@ module.exports.removeFile = async ({ Id, Del }) => {
       message.msg = "Server Error 500";
     }
   } else {
-    message.msg = "File not found";
+    message.msg = "File not found on db";
   }
   socket.emit("file-removed", message);
+};
+
+const getCoverPath = (name) => {
+  return path.join(process.cwd(), "images", "covers", "folder", name + ".jpg");
+};
+
+module.exports.renameFolder = async ({ Id, Name }) => {
+  let folder = await db.folder.findOne({
+    where: { Id },
+    include: { model: db.directory },
+  });
+  let success = false;
+  let msg = "Folder not found on system";
+  if (folder) {
+    try {
+      let basePath = folder.Directory.FullPath;
+
+      const oldPath = path.join(basePath, folder.Name);
+      let FullPath = path.join(basePath, Name);
+      if (fs.existsSync(oldPath)) {
+        fs.moveSync(oldPath, FullPath);
+        msg = "Folder Rename Successfully";
+        db.file.update({ FullPath }, { where: { FolderId: Id } });
+
+        let oldCover = getCoverPath(folder.Name);
+        if (fs.existsSync(oldCover)) fs.moveSync(oldCover, getCoverPath(Name));
+      }
+
+      await folder.update({ Name });
+      success = true;
+    } catch (err) {
+      console.log(err);
+    }
+    socket.emit("folder-renamed", { success, msg, Name });
+  }
+};
+
+module.exports.removeFolder = async (Id) => {
+  let folder = await db.folder.findByPk(Id);
+  let success = false;
+  if (folder) {
+    try {
+      await folder.destroy();
+      success = true;
+
+      let cPath = getCoverPath(folder.Name);
+      if (fs.existsSync(cPath)) fs.removeSync(cPath);
+    } catch (err) {
+      console.log(err);
+    }
+  }
+  socket.emit("folder-removed", { success });
 };
